test/cqlpy: move Vector Store mock into a shared module
The ANN mock defined inline in `test_vector_search_with_vector_store_mock.py`
will soon be needed by FTS tests too. Move it verbatim into a shared module
so both suites can run against a single HTTP server instance.

Fixtures moved to `conftest.py` so pytest auto-discovers them without
any explicit import in test files.
